backend/api/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
django-cors-headers
boto3
psycopg2-binary
gunicorn
uvicorn
whitenoise
psycopg2-binary

